### 2020-01-07 #12 @rougier
[Rp] Loupe
**Original article:** [N. P. Rougier. “LOUPE.” In: Tremplin Micro 19 (Mar. 1988), pp. 60–61](https://archive.org/details/tremplin_micro_newsletter_issue_19)

**PDF:**  https://github.com/rougier/TYRC-apple/blob/master/article/article.pdf
Metadata:  https://github.com/rougier/TYRC-apple/blob/master/article/metadata.yaml
Code:  https://github.com/rougier/TYRC-apple

Scientific domain:  None
Programming language:  Applesoft basic 😄
Suggested editor:  @khinsen
Suggested reviewer: @BrunoLevy
