### 2020-01-07 #12 @khinsen
Thanks for your submission! This has a good chance of remaining the oldest code submitted to the challenge.

@BrunoLevy, can you review this contribution?
