### 2020-01-07 #12 @BrunoLevy
Yep, I can do that (and I can test it on my Apple2 that is still in
working condition :-)

-- B
