### 2020-06-03 #12 @khinsen
@rougier Good point, I will try. I have a hard time with GitHub notifications myself, about 1/3 never arrive by mail and I stopped checking them on the site because of overload.
