### 2020-06-03 #12 @BrunoLevy
I have sent my review to Konrad (a while ago, and again today), please tell me if I need to post it / git commit it or something somewhere else.
Cheers,
-- Bruno
