### 2020-06-03 #12 @BrunoLevy
Hi Konrad, here it is ! (and I have inserted in it a few Apple][ photos and stories for fun).

I had a lot of fun reviewing this ReScience submission, because I have an Apple2 in working condition, and interestingly, it is an even older model (Apple][+) as compared to the ones used in the submission (Apple][e and Aplle//c).

The initial article, published in a French hobbyist magazine (Tremplin Micro) by the author when he was a teenager, loads a high resolution image from a floppy disk, lets the user choose a zone on the image, and displays a zoom of the image. The program is written in AppleSoft BASIC, a very popular language in the 80s (and mostly the only one available on the Apple without buying an expansive "language card").

The reproducibility report is very complete, and discribes the author's effort to re-understand 80's programming standards. In particular, the program involves several advanced programming techniques, such as the built-in command for vector graphics encoded shapes (XDRAW), that requires writing a sequence of values that encode the shape directly in memory. Another advanced technique involves using BLOAD / BSAVE to directly load/save a block of memory to/from the floppy, which involves emembering the address of video RAM.

All these technicalities are well described in the reproducibility report. It could be made even more complete with explanations of the binary format used by XDRAW, I have it handy in the "AppleSoft Programmer's manual" (believe it or not, it is still in my office near the Apple ][). Ref: BASIC Programming Reference Manual, Applesoft, Page 92 "how to create a shape table"

There is an online copy [here](http://www.classiccmp.org/cini/pdf/Apple/AppleSoft%20II%20Basic%20Programming%20Manual.PDF)

Something else that may help is a tutorial about how to connect an AppleII to a modern machine and transfer disk images to them:

you will need:
- an AppleII SuperSerial Card (bought one on e-bay)
- a DB25 to DB9 cable
- a DB9 to USB cable

A couple of tweets that show the super serial board installation:
- [step 1](https://twitter.com/BrunoLevy01/status/1216768059576410112)
- [step 2](https://twitter.com/BrunoLevy01/status/1216791874469494785)
- [step 3](https://twitter.com/BrunoLevy01/status/1216792642337562632)
- [the wires](https://twitter.com/BrunoLevy01/status/1218602155013296135)

Follow instructions from [this website](https://adtpro.com/) (especially how to configure the super serial). Then you are ready to send disk images

Note: finding floppies in working condition may be hard. Nearly all of mine have some kind of mold inside, but finally found one of them that was OK [See this photo on twitter](https://twitter.com/BrunoLevy01/status/1227574171862999041) (X indicates read/write error ! Finally found one floppy that did not generate any X).

Once I transfered the disk image attached with the submission, I was able to start the program from the floppy. I had a small issue with the greeting program, that displays a message with small caps that my AppleII cannot display (my 1977 AppleII+ can only display 40 columns with uppercase characters !), but by looking at what was on the disk (CATALOG) I found the program to start (run LOUPE) and everything ran smoothly (at 1980's speed).

To conclude, although it is not really a scientific article, I think this submission is a "must see" for the reproducibility challenge, it will be fun to see, and, I bet, of interest for the participants.

I've got several photos / videos that I'll be glad to share with the author for the presentation if he wants. [Little video on Twitter showing N. Rougier's program running](https://twitter.com/BrunoLevy01/status/1230951112783560704)

Here is a couple of Apple][ photos.

![apple2](https://user-images.githubusercontent.com/19478253/83685364-d69f9600-a5e8-11ea-89fc-9a17cb8ccdbc.jpg)
The apple][, sitting on my desk at Inria, following me since 1979 !

![Apple2_repair](https://user-images.githubusercontent.com/19478253/83685375-dbfce080-a5e8-11ea-88da-420561ac681f.png)
A few years ago, a capacitor exploded in the power supply unit, releasing a lot of white smoke ! Fortunately, 80's technology is easy to repair !

![Apple2ASM](https://user-images.githubusercontent.com/19478253/83685392-e0c19480-a5e8-11ea-9609-3a5742239b48.jpg)
Programming the Apple][ in assembly, assembling by hand on paper ! Where is the MUL unstruction ? No MUL instruction, you need to implement it on your own, using shifts and adds, and with only three 8-bit registers, it is an adventure !

What's the point ? Well, I guess in the ReScience galaxy, everybody knows what's the point, it is cool, and good science is reproducible science.
There is another benefit: when you design [your own RISC-V processor](https://twitter.com/BrunoLevy01/status/1265725361918226432), playing with old hardware helps a lot understanding what's going on (and funny enough, this one does not have MUL either, it is implemented in software !).

Best wishes,

-- B

P.S. Would be fun to write a faster version in 6502 assembly (we could organize a retrocomputing hackaton :-)
P.P.S. More Apple][ resources:
- [ROM disassembly](https://downloads.reactivemicro.com/Users/David_Craig/Apple2OriginalROMInfo.TXT) (Oh my god, if I only had this in 1984 !!)
- [InCider magazine archives](https://apple2online.com/index.php?p=1_70_inCider-Magazine)

END OF LINE.
