### 2020-06-03 #12 @BrunoLevy
Done ! (and I have inserted a couple of photos / links to tweets that
may be interesting)

Cheers,

-- B
