### 2020-06-04 #12 @khinsen
Thanks @BrunoLevy for this detailed, informative, and entertaining review. My provisional conclusion is that we should perhaps encourage the use of Apple ][s in scientific computing for improving the quality of everything.

I also conclude that @rougier's submission is accepted! I'll take care of the publishing steps today or tomorrow.
