### 2020-06-04 #12 @khinsen
@rougier Could you please generate the final PDF? You seem to have used an old version of the template, since it lacks the Makefile. Updated metadata is at https://github.com/rougier/TYRC-apple/pull/1.
