### 2020-06-09 #12 @rougier
Many thanks for the review @BrunoLevy. I've seen most of your photos on twitter but did not see all of them. Thanks again. I'm +1 for a retro hackaton. In fact, I would like to write a (non trivial) scientific result as if we were in 1980 but with out 2020 knowledge and try to guarantee this old-new result is reproducible from 1980 through 2020 (bit for bit). Just to check it is possible.
