### 2020-06-09 #12 @rougier
Thanks @khinsen, I'll update the paper.
