### 2020-06-09 #12 @rougier
@khinsen I've merged your update, I would need the DOI once you have it.
