### 2020-06-09 #12 @khinsen
@rougier DOI is here: https://github.com/rougier/TYRC-apple/pull/2
