### 2020-06-15 ae639e0
Regenerate metadata.tex and rebuild article
