+++ article/metadata.tex
@@ -4,15 +4,15 @@
 \def \codeDOI{}
 \def \dataURL{}
 \def \dataDOI{}
-\def \editorNAME{}
-\def \editorORCID{}
-\def \reviewerINAME{}
+\def \editorNAME{Konrad Hinsen}
+\def \editorORCID{0000-0003-0330-9428}
+\def \reviewerINAME{Bruno Levy}
 \def \reviewerIORCID{}
 \def \reviewerIINAME{}
 \def \reviewerIIORCID{}
 \def \dateRECEIVED{01 January 2020}
-\def \dateACCEPTED{}
-\def \datePUBLISHED{}
+\def \dateACCEPTED{04 June 2020}
+\def \datePUBLISHED{09 June 2020}
 \def \articleTITLE{[Rp] LOUPE}
 \def \articleTYPE{Reproduction}
 \def \articleDOMAIN{Ten Years Reproducibility Challenge}
@@ -26,12 +26,12 @@
 \def \replicationDOI{}
 \def \contactNAME{Nicolas P. Rougier}
 \def \contactEMAIL{Nicolas.Rougier@inria.fr}
-\def \articleKEYWORDS{ten years reproducibility challenge, apple 2e}
+\def \articleKEYWORDS{ten years reproducibility challenge, apple //e}
 \def \journalNAME{ReScience C}
-\def \journalVOLUME{5}
-\def \journalISSUE{2}
-\def \articleNUMBER{}
-\def \articleDOI{}
+\def \journalVOLUME{6}
+\def \journalISSUE{1}
+\def \articleNUMBER{3}
+\def \articleDOI{10.5281/zenodo.3886628}
 \def \authorsFULL{Nicolas P. Rougier}
 \def \authorsABBRV{N.P. Rougier}
 \def \authorsSHORT{Rougier}
