### 2020-06-15 #12 @khinsen
Thanks @rougier!

The article has been published (http://doi.org/10.5281/zenodo.3886628) and will soon be listed on the ReScience C Web site.
